Conclude rainbow mode
Found a way to activate the generation of a new color everytime a
mouseover event is triggered. The sketch is still lacking the shadow
mode which will be able to gradually apply 10% alphablack on each
mouseover event. Now that I understand the process correctly I believe
it will be easier to implement.