Add shade function.
I found a way to create a shade like effect working with rgb values.
by taking these 3 values and subtracting 10% of the max rgb value (25),
the color will get increasingly dark, but it's still not as accurate as
it would be by running a conversion to HSL, which is a color value
that takes lighting as it's own value. That would allow us to simply
subtract 10 from the current light value and get a perfect shading
effect. Right now, tho, that requires knowledge I believe I'm far
from accquiring (I might be wrong) so I'll leave it as it is for now.

This is a project I'd definitely like to revisit in the future, but for
now I'm incredibly happy with the progress I've made knowledgewise
while working the many problems that emerged.